metrics: Stabilize `num_active_tasks`
1. stabilize `num_active_tasks`
2. Rename internal counters to match `num_active_tasks`
3. Split `rt_metrics` into `rt_metrics` and `rt_unstable_metrics`.
