+++ tokio/src/runtime/scheduler/mod.rs
@@ -173,6 +173,10 @@ cfg_rt! {
                 Handle::MultiThreadAlt(handle) => handle.num_workers(),
             }
         }
+
+        pub(crate) fn num_active_tasks(&self) -> usize {
+            match_flavor!(self, Handle(handle) => handle.num_active_tasks())
+        }
     }
 
     cfg_unstable_metrics! {
@@ -191,10 +195,6 @@ cfg_rt! {
 
             pub(crate) fn num_idle_blocking_threads(&self) -> usize {
                 match_flavor!(self, Handle(handle) => handle.num_idle_blocking_threads())
-            }
-
-            pub(crate) fn alive_tasks_count(&self) -> usize {
-                match_flavor!(self, Handle(handle) => handle.alive_tasks_count())
             }
 
             pub(crate) fn scheduler_metrics(&self) -> &SchedulerMetrics {
