+++ tokio/src/runtime/scheduler/multi_thread/handle/metrics.rs
@@ -9,6 +9,10 @@ impl Handle {
         self.shared.worker_metrics.len()
     }
 
+    pub(crate) fn num_active_tasks(&self) -> usize {
+        self.shared.owned.num_active_tasks()
+    }
+
     cfg_unstable_metrics! {
         cfg_64bit_metrics! {
             pub(crate) fn spawned_tasks_count(&self) -> u64 {
@@ -25,10 +29,6 @@ impl Handle {
 
         pub(crate) fn num_idle_blocking_threads(&self) -> usize {
             self.blocking_spawner.num_idle_threads()
-        }
-
-        pub(crate) fn alive_tasks_count(&self) -> usize {
-            self.shared.owned.alive_tasks_count()
         }
 
         pub(crate) fn scheduler_metrics(&self) -> &SchedulerMetrics {
